refactor(menubar): extraer AppMenubar.vue y mover quick-add al slot #end
- Nuevo AppMenubar.vue: encapsula Menubar, slot #item con <a href>, slot #end
  con botón + para Super Admin y SplitButton de configuración
- MenubarQuickAdd.vue: elimina FAB flotante, expone open() via defineExpose
  para ser controlado desde AppMenubar via ref
- AppLayout.vue: sustituye bloque Menubar + lógica por <AppMenubar />
